Scene1 is   a test scene
PS：use this Scene1 to work on the game mechanics and introduce the general vibe of the game/environment.

I put all the packaged files in it (just in this folder, try Build and Run in Unity for the startup method)...because I really don't know where to put it.