Added PlayerInteract.cs for Object Interaction
I have added the PlayerInteract.cs script to the project. This script allows the player to interact with objects in the game world by casting a ray and triggering interactions using a specified key. The script also includes functionality to display a debug ray for debugging purposes and the potential to update the UI with prompt messages.